+++ sdk/src/client/bindings/python/examples/build_foundry.py
@@ -11,7 +11,7 @@ token_scheme = TokenScheme(0, 32, 64)
 unlock_conditions = [
     UnlockCondition(
         UnlockConditionType.ImmutableAliasAddress,
-        Address(AddressType.ALIAS, '0xa5c28d5baa951de05e375fb19134ea51a918f03acc2d0cee011a42b298d3effa')
+        AliasAddress('0xa5c28d5baa951de05e375fb19134ea51a918f03acc2d0cee011a42b298d3effa')
     )
 ]